embed operation in bridge and processor
