+++ manager-parent/manager-core/src/main/resources/db/migration/0.2/V0.2.1__create_v2_db.sql
@@ -1,57 +1,53 @@
 CREATE TABLE BRIDGE_V2 (
-    id                 varchar(255) NOT NULL PRIMARY KEY,
+    id                             varchar(255) NOT NULL PRIMARY KEY,
 
-    customer_id        varchar(255) NOT NULL,
-    organisation_id    varchar(255) NOT NULL,
-    owner              varchar(255) NOT NULL,
-    subscription_id    varchar(255) NOT NULL,
+    customer_id                    varchar(255) NOT NULL,
+    organisation_id                varchar(255) NOT NULL,
+    owner                          varchar(255) NOT NULL,
+    subscription_id                varchar(255) NOT NULL,
 
-    name               varchar(255) NOT NULL,
-    cloud_provider     varchar(255) NOT NULL,
-    region             varchar(255) NOT NULL,
+    name                           varchar(255) NOT NULL,
+    cloud_provider                 varchar(255) NOT NULL,
+    region                         varchar(255) NOT NULL,
 
-    endpoint           varchar(255),
-    shard_id           varchar(255) NOT NULL,
+    endpoint                       varchar(255),
+    shard_id                       varchar(255) NOT NULL,
 
-    submitted_at       timestamp    NOT NULL,
-    published_at       timestamp,
+    submitted_at                   timestamp    NOT NULL,
+    published_at                   timestamp,
 
-    version            integer      NOT NULL default 0,
-    generation         integer      NOT NULL default 0,
+    operation_type                 varchar(255) NOT NULL,
+    operation_requested_at         timestamp    NOT NULL,
+
+    version                        integer      NOT NULL default 0,
+    generation                     integer      NOT NULL default 0,
 
     unique (customer_id, name)
 );
 
 
 CREATE TABLE PROCESSOR_V2 (
-    id                 varchar(255) NOT NULL PRIMARY KEY,
+    id                             varchar(255) NOT NULL PRIMARY KEY,
 
-    name               varchar(255) NOT NULL,
-    flows              jsonb        NOT NULL,
+    name                           varchar(255) NOT NULL,
+    flows                          jsonb        NOT NULL,
 
-    owner              varchar(255) NOT NULL,
+    owner                          varchar(255) NOT NULL,
 
-    submitted_at       timestamp    NOT NULL,
-    published_at       timestamp,
+    submitted_at                   timestamp    NOT NULL,
+    published_at                   timestamp,
 
-    version            integer      NOT NULL default 0,
-    generation         integer      NOT NULL default 0,
+    version                        integer      NOT NULL default 0,
+    generation                     integer      NOT NULL default 0,
 
-    bridge_id          varchar(255) NOT NULL,
+    bridge_id                      varchar(255) NOT NULL,
+
+    operation_type                 varchar(255) NOT NULL,
+    operation_requested_at         timestamp    NOT NULL,
 
     unique (bridge_id, name),
     constraint fk_bridge foreign key (bridge_id) references BRIDGE_V2 (id)
 );
-
-CREATE TABLE OPERATION (
-    id                   varchar(255)   NOT NULL PRIMARY KEY,
-    type                 varchar(255)   NOT NULL,
-    requested_at         timestamp      NOT NULL,
-
-    managed_resource_id  varchar(255)
-);
-
-CREATE INDEX managed_resource_id_index ON OPERATION (managed_resource_id);
 
 CREATE TABLE CONDITION (
     id                     varchar(255) NOT NULL PRIMARY KEY,
@@ -63,7 +59,7 @@ CREATE TABLE CONDITION (
     errorCode              varchar(255),
     component              varchar(255) NOT NULL,
 
-    operation_id           varchar(255) NOT NULL,
-
-    constraint fk_operation foreign key (operation_id) references OPERATION (id)
+    managed_resource_id    varchar(255) NOT NULL
 );
+
+CREATE INDEX managed_resource_id_index ON CONDITION (managed_resource_id);
