--- manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/persistence/dao/OperationDAO.java
@@ -1,14 +1,0 @@
-package com.redhat.service.smartevents.manager.v2.persistence.dao;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.transaction.Transactional;
-
-import com.redhat.service.smartevents.manager.v2.persistence.models.Operation;
-
-import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
-
-@ApplicationScoped
-@Transactional
-public class OperationDAO implements PanacheRepositoryBase<Operation, String> {
-    // Empty atm
-}
