+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/persistence/models/Condition.java
@@ -8,10 +8,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import com.redhat.service.smartevents.infra.v2.api.models.ComponentType;
@@ -42,9 +39,8 @@ public class Condition {
     @Enumerated(EnumType.STRING)
     private ComponentType component;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "operation_id")
-    private Operation operation;
+    @Column(name = "managed_resource_Id")
+    private String managedResourceId;
 
     @Column(name = "last_transition_time", columnDefinition = "TIMESTAMP", nullable = false)
     private ZonedDateTime lastTransitionTime;
@@ -101,12 +97,12 @@ public class Condition {
         this.component = component;
     }
 
-    public Operation getOperation() {
-        return operation;
+    public String getManagedResourceId() {
+        return managedResourceId;
     }
 
-    public void setOperation(Operation operation) {
-        this.operation = operation;
+    public void setManagedResourceId(String managedResourceId) {
+        this.managedResourceId = managedResourceId;
     }
 
     public ZonedDateTime getLastTransitionTime() {
