+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/persistence/models/ManagedResourceV2.java
@@ -1,0 +1,32 @@
+package com.redhat.service.smartevents.manager.v2.persistence.models;
+
+import javax.persistence.Column;
+import javax.persistence.Embedded;
+import javax.persistence.MappedSuperclass;
+
+import com.redhat.service.smartevents.manager.core.models.ManagedResource;
+
+@MappedSuperclass
+public class ManagedResourceV2 extends ManagedResource {
+    @Column(name = "owner", nullable = false, updatable = false)
+    protected String owner;
+
+    @Embedded
+    protected Operation operation;
+
+    public String getOwner() {
+        return owner;
+    }
+
+    public void setOwner(String owner) {
+        this.owner = owner;
+    }
+
+    public Operation getOperation() {
+        return operation;
+    }
+
+    public void setOperation(Operation operation) {
+        this.operation = operation;
+    }
+}
