+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/persistence/models/Operation.java
@@ -1,44 +1,23 @@
 package com.redhat.service.smartevents.manager.v2.persistence.models;
 
 import java.time.ZonedDateTime;
-import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
 
 import javax.persistence.Column;
-import javax.persistence.Entity;
+import javax.persistence.Embeddable;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
 
 import com.redhat.service.smartevents.infra.v2.api.models.OperationType;
 
-@Entity
-@Table(name = "OPERATION")
+@Embeddable
 public class Operation {
 
-    @Id
-    private String id = UUID.randomUUID().toString();
-
-    @Column(name = "type", nullable = false)
+    @Column(name = "operation_type", nullable = false)
     @Enumerated(EnumType.STRING)
     private OperationType type;
 
-    @Column(name = "requested_at", columnDefinition = "TIMESTAMP", nullable = false)
+    @Column(name = "operation_requested_at", columnDefinition = "TIMESTAMP", nullable = false)
     private ZonedDateTime requestedAt;
-
-    @Column(name = "managed_resource_id", nullable = false)
-    private String managedResourceId;
-
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "operation")
-    private Set<Condition> conditions;
-
-    public String getId() {
-        return id;
-    }
 
     public OperationType getType() {
         return type;
@@ -56,40 +35,4 @@ public class Operation {
         this.requestedAt = requestedAt;
     }
 
-    public String getManagedResourceId() {
-        return managedResourceId;
-    }
-
-    public void setManagedResourceId(String managedResourceId) {
-        this.managedResourceId = managedResourceId;
-    }
-
-    public Set<Condition> getConditions() {
-        return conditions;
-    }
-
-    public void setConditions(Set<Condition> conditions) {
-        this.conditions = conditions;
-    }
-
-    /*
-     * See: https://vladmihalcea.com/how-to-implement-equals-and-hashcode-using-the-jpa-entity-identifier/
-     * In the context of JPA equality, our id is our unique business key as we generate it via UUID.
-     */
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Operation operation = (Operation) o;
-        return id.equals(operation.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
 }
