+++ manager-parent/manager-v2/src/test/java/com/redhat/service/smartevents/manager/v2/persistence/dao/BridgeDAOTest.java
@@ -39,5 +39,8 @@ public class BridgeDAOTest {
 
         Bridge retrieved = bridgeDAO.findById(bridge.getId());
         assertThat(retrieved.getId()).isEqualTo(bridge.getId());
+        assertThat(retrieved.getOperation()).isNotNull();
+        assertThat(retrieved.getOperation().getType()).isNotNull();
+        assertThat(retrieved.getOperation().getRequestedAt()).isNotNull();
     }
 }
