+++ manager-parent/manager-v2/src/test/java/com/redhat/service/smartevents/manager/v2/persistence/dao/ConditionDAOTest.java
@@ -6,17 +6,16 @@ import javax.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import com.redhat.service.smartevents.manager.v2.persistence.models.Bridge;
 import com.redhat.service.smartevents.manager.v2.persistence.models.Condition;
-import com.redhat.service.smartevents.manager.v2.persistence.models.Operation;
 import com.redhat.service.smartevents.manager.v2.utils.DatabaseManagerUtils;
 import com.redhat.service.smartevents.test.resource.PostgresResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import static com.redhat.service.smartevents.manager.v2.TestConstants.DEFAULT_BRIDGE_ID;
+import static com.redhat.service.smartevents.manager.v2.utils.Fixtures.createBridge;
 import static com.redhat.service.smartevents.manager.v2.utils.Fixtures.createCondition;
-import static com.redhat.service.smartevents.manager.v2.utils.Fixtures.createOperation;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @QuarkusTest
@@ -25,9 +24,6 @@ public class ConditionDAOTest {
 
     @Inject
     ConditionDAO conditionDAO;
-
-    @Inject
-    OperationDAO operationDAO;
 
     @Inject
     DatabaseManagerUtils databaseManagerUtils;
@@ -41,15 +37,13 @@ public class ConditionDAOTest {
     @Transactional
     public void testStoreOperation() {
 
-        Operation operation = createOperation(DEFAULT_BRIDGE_ID);
-        Condition condition = createCondition(operation);
-        condition.setOperation(operation);
+        Bridge bridge = createBridge();
+        Condition condition = createCondition(bridge);
 
-        operationDAO.persist(operation);
         conditionDAO.persist(condition);
 
         Condition retrieved = conditionDAO.findById(condition.getId());
         assertThat(retrieved.getId()).isEqualTo(condition.getId());
-        assertThat(retrieved.getOperation().getId()).isEqualTo(operation.getId());
+        assertThat(retrieved.getManagedResourceId()).isEqualTo(bridge.getId());
     }
 }
